refactor: reorganize route imports
