+++ app/admin/reports/page.tsx
@@ -8,7 +8,7 @@ import { Anchor, Box, Center, Container, Title } from '@mantine/core'
 import dayjs from 'dayjs'
 import { fetchReportList } from '@/utils/fetch'
 import { InaccuracyReportListDTO } from '@/app/types'
-import { barometerRoute } from '@/app/constants'
+import { barometerRoute } from '@/utils/routes-front'
 import { Table } from '@/app/components/table'
 import { Pagination } from '@/app/components/pagination'
