+++ app/collection/items/[slug]/components/edit-fields/images-edit.tsx
@@ -27,7 +27,8 @@ import { useEffect, useMemo, useState } from 'react'
 import { useDisclosure } from '@mantine/hooks'
 import { BarometerDTO } from '@/app/types'
 import sx from './styles.module.scss'
-import { barometerRoute, googleStorageImagesFolder } from '@/app/constants'
+import { googleStorageImagesFolder } from '@/utils/constants'
+import { barometerRoute } from '@/utils/routes-front'
 import { showError, showInfo } from '@/utils/notification'
 import { createImageUrls, deleteImage, updateBarometer, uploadFileToCloud } from '@/utils/fetch'
 import { getThumbnailBase64 } from '@/utils/misc'
