+++ app/components/header/mobile-menu/mobile-menu.tsx
@@ -20,8 +20,9 @@ import { SiMaildotru, SiInstagram } from 'react-icons/si'
 import { IoIosArrowForward as Arrow } from 'react-icons/io'
 import { useSession } from 'next-auth/react'
 import { AccessRole } from '@prisma/client'
-import { instagram, email, categoriesRoute } from '@/app/constants'
-import { menuData } from '../menudata'
+import { instagram, email } from '@/utils/constants'
+import { categoriesRoute } from '@/utils/routes-front'
+import { menuData } from '@/utils/menudata'
 import { useBarometers } from '@/app/hooks/useBarometers'
 import { isAdmin } from '../../is-admin'
