--- app/constants.ts
@@ -1,25 +1,0 @@
-// Social networks, contacts
-export const email = 'post@barometers.info'
-export const instagram = 'https://www.instagram.com/barometers_realm/'
-export const twitterAccount = '@barometer_realm'
-// API Routes
-const base = '/api/v2/'
-export const barometersApiRoute = `${base}barometers/`
-export const categoriesApiRoute = `${base}categories/`
-export const manufacturersApiRoute = `${base}manufacturers/`
-export const conditionsApiRoute = `${base}conditions/`
-export const imageUploadApiRoute = `${base}upload/images/`
-export const barometersSearchRoute = `${base}search/`
-export const reportRoute = `${base}report/`
-// Frontend routes
-export const addBarometerRoute = '/admin/add-barometer/'
-export const viewReportsRoute = '/admin/reports/'
-export const categoriesRoute = '/collection/categories/'
-export const barometerRoute = '/collection/items/'
-export const newArrivals = '/collection/new-arrivals/'
-export const manufacturersRoute = '/manufacturers/'
-// External links
-export const googleStorageImagesFolder = 'https://storage.googleapis.com/barometers/'
-export const github = 'https://github.com/shenshin'
-
-export const BAROMETERS_PER_CATEGORY_PAGE = 12
