+++ app/page.tsx
@@ -1,7 +1,7 @@
 import { Container, Grid, GridCol, Group } from '@mantine/core'
 import { HeadingImage } from './components/heading-image'
 import { CategoryCard } from './components/category-card'
-import { categoriesRoute } from './constants'
+import { categoriesRoute } from '@/utils/routes-front'
 import { SearchField } from './components/search-field'
 import { getCategories } from './api/v2/categories/getters'
 import { NewArrivals } from './components/new-arrivals'
