+++ utils/constants.ts
@@ -1,0 +1,10 @@
+// Social networks, contacts
+export const email = 'post@barometers.info'
+export const instagram = 'https://www.instagram.com/barometers_realm/'
+export const twitterAccount = '@barometer_realm'
+
+// External links
+export const googleStorageImagesFolder = 'https://storage.googleapis.com/barometers/'
+export const github = 'https://github.com/shenshin'
+
+export const BAROMETERS_PER_CATEGORY_PAGE = 12
