+++ utils/routes-back.ts
@@ -1,0 +1,9 @@
+// API Routes
+const base = '/api/v2/'
+export const barometersApiRoute = `${base}barometers/`
+export const categoriesApiRoute = `${base}categories/`
+export const manufacturersApiRoute = `${base}manufacturers/`
+export const conditionsApiRoute = `${base}conditions/`
+export const imageUploadApiRoute = `${base}upload/images/`
+export const barometersSearchRoute = `${base}search/`
+export const reportRoute = `${base}report/`
