+++ utils/routes-front.ts
@@ -1,0 +1,15 @@
+// Frontend routes
+export const homeRoute = '/'
+export const historyRoute = '/history/'
+export const aboutRoute = '/about/'
+export const brandsRoute = '/brands/'
+
+// Collection
+export const categoriesRoute = '/collection/categories/'
+export const barometerRoute = '/collection/items/'
+export const newArrivals = '/collection/new-arrivals/'
+
+// Admin
+export const adminRoute = '/admin/'
+export const addBarometerRoute = '/admin/add-barometer/'
+export const viewReportsRoute = '/admin/reports/'
